requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
mysql-connector==2.2.9
numpy==1.17.1
pandas==0.25.1
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.7
